feat: improve main site with server components and responsive fixes
Main Page Improvements:
- Add missing Image import for Next.js optimization
- Maintain existing responsive layouts and functionality

Post Page Architecture:
- Convert post/[slug]/page.tsx to Server Component for better SEO
- Add comprehensive metadata generation with OpenGraph and Twitter cards
- Extract client-side functionality to PostViewClient.tsx
- Improve server-side rendering performance

PostViewClient Component:
- Maintain all existing interactive functionality
- Fix Image component usage with proper width/height props
- Improve responsive image handling
- Better error handling for missing images

SEO & Performance Benefits:
- Server-side metadata generation for better social sharing
- Improved Core Web Vitals with proper image optimization
- Better search engine indexing with server-rendered content
- Maintained client-side interactivity where needed
